Remove actions that try to remove non-existing devices (#906807)
When clicking in the partitioning GUI it may happen that an action trying
to destroy a device that doesn't exist in the list ends up in the list of the
actions to be processed. Those cases are bugs and it should never happen, but
instead of doing changes in partitioning code we can simply remove such invalid
actions in the pruneActions() method.