Fix behavior (and docs) for ks=nfs:<path>/ (#1094645)
The loader-era behavior for ks=nfs:...:/path/ was to append
"<ip>-kickstart" as a filename.

In the loader rewrite, the behavior was accidentally changed so this
only happened when you did the magical automatic kickstart detection via
plain "ks". And this is how it was documented.

Bug #910550 pointed out that this doesn't match the F16 (i.e. RHEL6)
behavior. I read the old code and we concluded that we should add that
for *any* URL ending in a /, and so commit c85dd38 "fixed" the problem.

Turns out I was wrong, though! Testing confirms that RHEL6 behavior is
to add "<ip>-kickstart" to NFS URLs (and *only* NFS urls) ending in '/'.

This patch fixes that behavior, and corrects the documentation.