Return CLI help text and version without delay
If the CLI option help is requested with the --help/-h or --version switch,
return the help text or version as quickly as possible and without unrelated
output.

Previously Anaconda took a couple of seconds to respond and the user
was greeted by unrelated status and error messages.

Also do basic logging setup when short circuiting the normal Anaconda init,
this gets rid of the annoying "no handlers found" error message on stdout
when printing the Anaconda help or version.

Related: rhbz:#1088459