Correct matching of md raid names containing special characters (#1096889)
In md raid device names, udev replaces spaces and tabs with '_' and '/'
with '-'. Such names are returned by udev_device_get_md_name(), but in
DeviceTree.isIgnored() this is compared with names from
DeviceTree.exclusiveDisks, which are in original form. Because of this,
md raid devices with such names are not found in exclusiveDisks.

This patch adds substitution of special characters when comparing
md_name with names from exlusiveDisks.

Resolves: rhbz#1096889

Signed-off-by: Samantha N. Bueno <sbueno+anaconda@redhat.com>