rpmostreepayload: Use systemd-tmpfiles rather than handrolling mkdir
I hit an issue where something in Anaconda may invoke a program or
library which uses rpm, and because the OSTree model has /var/lib/rpm
-> /usr/share/rpm, but that's created during boot, I'd end up with a
new empty rpmdb in /var/lib.

Fix this by switching to running systemd-tmpfiles on the target root
*only* for the directories we'll need at install time.  For example,
we don't need /var/spool/abrt until the target is running, so don't
try to create it.

Resolves: rhbz#1113535
Port of rpmostreepayload patches from master
commit a5f48d26312bfde6800d8bb608c5f08186c8eed1