Better visually distinguish TUI spokes states
If a TUI spoke is mandatory and not completed, its status "icon" should be [!],
if it is not mandatory and is completed, it should be [x] and if it is not
mandatory and not completed, it should be [ ].

Otherwise e.g. the User spoke looks like completed even when just root password
is set (which makes the User spoke not mandatory) which might be confusing.