Fix the modify software tooltip on the installation options dialogs.
The tooltip indicating metadata was still being downloaded was not displaying.
This is because the link in the label's markup had no title and GTK really
wants a title.  So, give it one and remove the tooltip from the rest of the
label.

While I was at it, I changed _add_modify_watcher to just take a widget that
we already have instead of a name, and removed a pointless show_all.