Allow file:// url handler in --repo arg
When doing --image installations it would be nice to be able to point to
a local repo without needing a http server. The fix here is simple,
treat a --repo=file:/// as a url method and everything works.

There was a suggestion to allow pointing directly to an iso file, but
adding support for that would mean complicating the code. It is better
to allow the user to mount their image and point Anaconda to the
moutpoint.