Parse boot options before parsing CLI options (#1101341)
First parse all boot options relevant to Anaconda and only then
parse CLI options (if any). This makes it possible for CLI options
to override boot options, which could be useful for example
when running Anaconda during a live installation.

This also fixes issues with boot options that have a default value
in the Anaconda argument parser - such options were previously ignored,
resulting for example in the "text" boot option not triggering the text mode.