Fix issues with auto* and version.py
automake is really kind of terrible at python.

Filter the generated version.py out of the dist files so that builds
from a dist tarball don't end up with a both a $srcdir and $builddir
copy of version.py. Ignore a pylint error about importing version since
version.py ends up in a different directory than the rest of the
pyanaconda module.