allow /boot on btrfs subvol or filesystem
This reverses the patch disallow /boot on btrfs
which was done in response to BZ 864198.

With grub2-2.02 and os-prober-1.58-6 both in rawhide, two of the
problem packages for /boot on btrfs have been eliminated.

The remaining problem is grubby and a patch has been submitted to
implement btrfs support in grubby.c (see BZ 1094489).  To test the
updated grubby, btrfs support needs to be re-enabled in anaconda.

Re-enabling btrfs on /boot is not a big issue as that was the
situation prior to Fedora 20 and both grub2 and os-prober now support
btrfs which was not true in Fedora 20.