move new-kernel-pkg rpmposttrans to end of install
This address BZ 1099627 and BZ 1100504.

A regular DVD or network install creates the rescue kernel
and initramfs as part of the install performed by anaconda-yum
and not postInstall.

This needs to be done in Live also so that when the bootloader
configuration file is created, it has the correct files to work with.

Note the comment in the code that says it must be done before
bootloader.