always rescan for vmlinuz if rescueKernelList
This addresses rhbz1036349 and rhbz1100504

In order to create the rescue kernel/initramfs, kernelVersionList
is referenced creating and populating _kernelVersionList with
rescueKernelList null since the rescue kernel has not been created yet.

The solution is to always rescan /boot/vmlinuz-* if rescueKernelList
is called.  This is not really costly since it is only used once.

Note: Invalidating the cache when the recue kernel is created would
involve a lot more code which, of course, is also more potential for
bugs.