Fix up quoting around passing args to anaconda in the gui test.
We go through a lot of layers - command line to shell a couple different ways,
then through noseconfig-testconfig, and into a python template, and finally
written into a python file.  Along the way, the string that actually runs
liveinst ended up looking like so:

    "liveinst "--updates=http://wherever/updates.img""

Which clearly is never going to work.  The changes in this patch make that one
continuous string, plus handle what happens if you want to pass multiple
arguments or no arguments.

Also, remove the quotes from around the livecd variable.  I don't expect any
one to make a livecd image with spaces in it.  I know I'm not going to set
up automated testing to do that.