Reword the close button on the software spoke's error dialog.
If you hit a problem with your software selections (like file conflicts or
other packaging problems), you don't really want to cancel.  What does that
even mean in this context?  Likely, you want to close the dialog so you can
change software selections.  For instance, you might change from a desktop
to a minimal install so you at least have something installed.

Thus, change the wording on the button.  What the button actually does has
not changed.  While I was at it I moved the button to the right side (default
action goes on the right) and reworded the label to reflect all the buttons.