change default for grub2 save_entry to 0
The value that anaconda sets save_entry in /boot/grub2/grubenv is
not valid as there is no corresponding menuentry with that title.  The
only reason things work now is that if grub2 finds no match, it boots
the kernel specified in the first (top) menuentry.

grub2-set-default --help results in:
  MENU_ENTRY is a number, a menu item title or a menu item identifier

This patch changes the default to be "0".

This results in the value now being correct, produces an expected
result (booting the first kernel), and not requiring any changes
to grubby.

The user will still be able to use grub2-set-default to change the
default value to other settings.