Add a class BaseStandalone.
This class is an abstract class containing the shared parts of HubWindow
and StandaloneWindow. It provides quit-button and continue-button
properties, quit-clicked and continue-clicked signals emitted when the
buttons are pressed, and the sidebar sizing and drawing methods.

Glade files using StandaloneWindow do not require any change. Glade files
using HubWindow may need to be modified to connect the quit-button or
continue-button properties.

On the python side, this removes the need for the register_event_cb
proxy for quit and continue events. Connect GraphicalUserInterface
directly to the now-shared signals.