Add a window to manage Anaconda screen transitions.
Instead of each hub and spoke being a separate Gtk window, create a
single window to contain every non-dialog screen. This changes the
parent type of AnacondaBaseWindow to GtkBin instead of GtkWindow.

This change is mostly helpful in live environments where we have less
control over the window manager. Each spoke and hub appeared as a
separate window the in the window manager and it made things look kind
of weird, as well as opening up the possibility for some troublesome
window interactions.

Remove the separate GtkWindowGroup for exception windows, since there's
no longer a separate Gtk.main running for spokes.