Mark dracut-network for installation on mpath. (#1007463)
This is probably a hack and not an ideal solution, however it is the
lesser of two evils AFAICT. Multipath SAN devices pop up as iSCSI
devices in syslog and even bear some iSCSI-related udev info, as well
as some FCoE-related udev info;  however, not enough to match as either
an iSCSI or FCoE device and trigger installation of the relevant
dracut-network package.

Didn't want to go messing with the udev functions since those look quite
fragile, so this seemed a safer option, since the devices are getting
detected as multipath devices at least.

Resolves: rhbz#1007463