Close yum objects before running post scripts (#1110148)
yum changed to using an exclusive lock on the history database file,
causing use of yum in %post to fail because anaconda wasn't closing out
access to yum before running the kickstart %post scripts.

complete now closes the ts objects, the ayum object and deletes the rpm
database locks.

Note that running yum from %post is not encouraged, it is better to
include the repositories and packages in the kickstart.

Resolves: rhbz#1110148