Make the configure mount point dialog taller. (#924182)
By default the ScrolledWindow in the configure mount point dialog is
tall enough for two and a half items, which is pretty much the minimum
number of items that can be in the dialog and have it make sense. Expand
it to the height that the container dialog uses, which is enough for
five or so items.