Remove the noipv6 Anaconda option
So it looks like noipv6 option is yet another option that Anaconda
parses but doesn't do anything with. The functionality that has been handled by
this option in the past seems to have moved to kernel (which uses the
same keyword) and we just check for "noipv6" being present on the boot
command line to apply our IPv6 tweaks.