Remove surprises from X startup.
As was noted in the comment, assuming that any SIGCHLD means that X
exited makes the startup process fragile, so stop doing that. Install a
global SIGCHLD handler to track exits from Xorg and metacity. metacity
failures will now throw an exception instead of being silently,
confusingly ignored. Add a fork wrapper to remove the race between
process startup and process monitoring.

Removed a comment about the importance of the window manager being the
first X connection. This is no longer true, which is fortunate as the
period between metacity starting and metacity connecting to the X server
presents an intractible race.