zRAM swap for Anaconda
zRAM is a standard block device like tmpfs with one nice difference -- its
content is compressed. One clever usage of such device is using it as a swap
device which effectively means one can have compressed RAM. Some experiments
show that by using zRAM as a swap device in the installation process anaconda
can install to machines with lower amount of RAM (succeeded with 320 MB) and
installations on low-RAM systems (< 1 GB) run faster due to usage of zRAM swap
instead/together with the standard swap we create and use for package
installation.

Specific limits, numbers and way of implementation may be changed in the future,
but this commit introduces the basic idea and straightforward approach.