Load icons by name instead of stock-id.
Rename the icons to the new, improved icon names used in
gnome-icon-theme.

gtk-apply has no equivalent in the brave new world with such icons in't.
For the refresh and dasdformat dialogs, use an uncolored checkbox,
emblem-ok-symbolic, instead. For the cart dialog, replace the pixbuf
renderer with a toggle renderer, which ends up getting the same symbolic
icon but in a way that they can't rename as easily.