Split kickstart arg handling
When inst.ks is used the Anaconda dracut module downloads it and
partially processes it, writing it to /run/install/ks.cfg so --ks
indicates a file at this location should be used.

When --kickstart <path> is used from the cmdline it means use that
kickstart file directly. No URL handling is done when this argument is
passed.