Set widgets to be focused when entering a spoke. (#1121285)
Defining the spokes and hubs in glade files without toplevels and then
putting them into a container breaks gtk's usual default focus behavior
along with the explicit has_focus that we had defined in the welcome
spoke. To work around gtk having 17 words for "focus" and yet not having
a way to say that the default thing in a container's child should grab
the focus, define in python which widgets to focus and grab the focus
for them during the GtkStack transitions.