Lessen the visible resize when entering the welcome and lang spokes
Set the two treeviews to take up the full width of the spoke instead of
being centered within the spoke and set the containing grid to
column-homogenous. This way the widths of the TreeViews are not
dependent upon the size of the strings in the ListStores. Always make
the language column scrollbar visible so the scrollbar won't appear
part of the way through the list populating. Switch the alignment
properties on the language text renderers so that the native string is
most of the way to the left and the english string is most of the way to
the right since I think that was what was intended anyway.

Realigned the text and rekajiggered the spacing to make things look less
weird with the guidance of duffy@redhat.com