Remove unused parts of the configure.ac files.
We never check the output of any of those program tests, so if you don't
have a working awk or grep installed, too bad. Nothing links directly
with libX11 anymore. We don't need an explicit check for pkg-config
because we always run PKG_CHECK_MODULES. nfs and ipv6 cannot be
disabled. We don't need to use glib directly. Caring about timezones has
been moved to an external library. There's no need to reset $LIBS
because none of the checks we have left set it in the first place.