Switch to kinder, gentler autoconf errors
The usual paradigm in autoconf is to fail configure if a dependency is
not found, but this is unhelpful in an environment that has 47 zillion
make targets that aren't directly connected to compling a program.
Instead of failing, print a message about missing dependencies at the
end of the ./configure with some ***'s, but still exit with a status 0
and let the Makefiles be generated. Any missing dependencies will cause
a failure during `make', and other targets like `make dist' will still
work.