Resurrect auditd
All of auditd except for the parts that call exit() was wrapped in
USESELINUX ifdefs, which was not defined, so our fake auditd hasn't
done anything for a while. Remove the ifdefs since building an auditd that
does that much nothing is kind of silly. Fix the LDFLAGS needed to link.
Switch from oom_adj to oom_score_adj since the former was deprecated
some time ago and complains when you use it now. Add auditd to
makeupdates.