Make storage sanity check aware of base RAM requirements (#1123466)
Minimum RAM requirements depend on the installation mode (GUI/TUI) so the
storage sanity check needs a parameter telling it what the current minimum RAM
requirement is.

Replace the use of the EARLY_SWAP_RAM constant (removed in the commit
0f200d705b6f167abc29cf93bdac13fa1bbaf6c4) with the NO_SWAP_EXTRA_RAM that has a
better name and gives the right information independent on tthe installation
mode.