Add a help button to every Anaconda screen
This makes context aware help possible for every Anaconda screen by
providing a help button in the upper right corner of the screen.
Clicking the button opens the yelp help viewer with either a help page
tailored for the given screen, the main page of the installation guide
(if no help is available for the screen) or a "no help found" placeholder
page.
The new helpFile GUIObject attribute is used to specify which help
file to show for the given screen. If no help file is specified
either the main page of the Installation Guide will be shown (if
available) or the placeholder page will be shown.