Ask users for enough space right at the first time (#876916)
If autopartitioning is not yet done because we are just about to ask user for
allocating enough space for us, we cannot determine the size of swap by summing
up the auto swap requests. Ask user for enough space for a swap we would like to
create if we had enough space instead.