Add noverifyssl and proxy support to dracut ks handling (#1116858)
When the kickstart is used to point to the location of the tree using
https it needs to also set rd.noverifyssl if the --noverifyssl flag is
set on the url install method on the kickstart.

Also add support for --proxy being passed to proxy cmdline option,
passed on to curl by dracut.

Resolves: rhbz#1116858