Re-enable addons as additional repositories.
This appears to have been broken due to a change in the format of the .treeinfo
file.  Addons are now listed under a variant heading, just like the main
installation source.  They are differentiated by type=addon, though.

While I was at it, make sure the repo editing box becomes sensitive when you
enable a repo.  It was not being made sensitive at all.

Resolves: rhbz#1061174