Handle spaces in inst.repo, kickstart nfs, and url commands (#1109933)
NFS paths and http urls may have spaces in them. When passed on the
cmdline they should be escaped and when in a kickstart the string should
be quoted.

This changes the anaconda dracut code so that kickstart nfs with spaces
is written using \x20 when output by parse-kickstart and the \x20 is
converted back to a space before mounting the nfs root. For http urls it
converts the spaces to %20 which can be passed through to curl without
any further conversion.

The reason for the conversion is that dracut loops over cmdline args
split by spaces, so they cannot be present in the cmdline or from the
output of parse-kickstart whis is parsed as if it were on the cmdline.

This also depends on a dracut change: mount_nfs needs to quote the
arguments passed to it and when it passes them on to nfs_to_vars.

Resolves: rhbz#1109933