Use one thread for payload setup.
Add an object to send events for payload states so that we can have one
copy of the package payload setup code instead of four. This removes
THREAD_PAYLOAD_MD and some duplicated code in THREAD_SOFTWARE_WATCHER,
which also removes the problems of THREAD_PAYLOAD_MD and
THREAD_SOFTWARE_WATCHER potentially accessing the yum environments and
groups properties simultaneously and causing a very confusing crash.

In the payload classes, remove the updateBaseRepo and gatherRepoMetadata
calls from setup so that the remainder of setup can be run separately
from these other two methods. For classes that no longer had any code in
setup, remove the method.

Resolves: rhbz#1125927