Fix installing from a second iso (#1040722)
This fixes handling of the source from a second iso. eg. boot from a
boot.iso with a DVD iso in a second drive.

MOUNT_DIR has been changed from /mnt/install to /run/install so that the
mount path will match /proc/mounts. Mounting via a symlink shows up as
the real path, not the symlink.

This was causing it to use util.unmount to teardown the install_device
instead of the device's teardown method, leaving it in a state where it
could later raise an error if the teardown method was used.

This splits out parts of the YumPayload reset method and streamlines the
setup of the source so that it unmounts the previous source, then sets
up the new one, gets releasever and resets the yum object without the
previous hack of having to call _setupInstallDevice twice.

Resolves: rhbz#1040722