Fix two problems with the volume label and combo on custom partitioning.
These are both caused by the fact that we reset the label's text away from
what is defined in glade.

(1) Keep the colon on the end of the label.

(2) Change the label's accelerator key to V, which does conflict with the remove
button but we can't use keyboard accelerators to activate those buttons anyway.