network: don't traceback on invalid team options (#1114282)
Resolves: rhbz#1114282

If team connection with invalid options is added in GUI (Connection Editor), NM
activates the connection which activates teamd which dies, and the team device
Anaconda is expecting to appear is not created so Anaconda tracebacks. The
patch lets Anaconda survive with just logging error message.  Quite a general
message, but we are not able to tell more than that for some reason the team
device was not created/found.

(cherry picked from commit 5566a7d4a4830c0ff1787ce7ff46265f5ea11fd9)
Related: rhbz#1196721