For yum-based installs, move the progress bar while packages are installing.
It's impossible to know how many packages will be installed before we start
doing stuff with the progress bar.  Just look at how realmd is handled to
understand that.  Thus, just move the bar along for every 10% of packages.
This also has the effect of making all the other steps of installation take
less of the total amount of the progress bar which makes package installation
look a little more serious.

Other payloads can gain this as-needed, though right now it looks like the
dnf payload would be the most critical.

Related: rhbz#1024403