Use first part of Product as repo name (#1128474)
Anaconda uses the Product from .buildstamp as the name of the repo to
use for mirror installations. This has to match an entry in
/etc/anaconda.repos.d/

We share the repositories between the various Products so we should only
use the first element (splitting on '-') as the name of the repo.

(cherry picked from commit 0d193bcb98cf61c407e88d016cab0015235a3f78)