fix inst.virtiolog (#1074499)
'inst.virtiolog' got lost in the loaderectomy, so we only ever used the
default port name.

This commit makes us use the port name given by 'inst.virtiolog=<name>'
if it's present. Otherwise we use the default name, as always.

Also, clarify and add a link for more info in boot-options.txt.