Add dialog box for adding DASDs.
This is the dialog box which pops up prompting for info so more (ECKD)
DASDs can be added.

Upon a successful add, users are returned to the filter. Due to the fact
that DASDs are detected as local storage, newly-added DASDs here are not
going to show up in the Search or System Z Devices tab. They do appear
under "Local Standard Disks" back in the main storage spoke though.