Revert productName repo name change (#1128474)
The products want the ability to make Product specific netinst isos.
Personally, I think this is the wrong approach, but in order to allow them the
flexibility I'll revert my .buildstamp Product patch.

This means that when making a compose the composer has to be sure they include
a package installing a repo to /etc/yum.repos.d/ that has an id matching the
lower-case version of the product name (eg. fedora-server or
fedora-workstation). Without this the closest mirror option will not work
because anaconda doesn't know the name of the repo to use.

(cherry picked from commit 51f204b9694400b512d0e8cf33c4b409ac0fa0d4)