Fix a problem where urllib2 is not getting pulled into the initrd.
python-six is too clever and quietly converts urllib imports into urllib2
imports, but in such a way that python-deps doesn't pick up on that fact.
Thus, urllib2 is not in the initrd and kickstart installs hit a traceback
when attempting to fetch the ks file.  The result is either a regular
interactive install or staring at a dracut prompt, depending on what else
you are doing.

There may be more of this in the future.