Fix q for quit issue in text mode (#997405)
Pressing 'q' to quit the installer in the ask VNC spoke of text mode was
not working. We can't rely on the global handling of this input,
unfortunately, since the installation environment is not set up at this
point. But, this should be a relatively simple way of adding the
functionality.

Resolves: rhbz#997405