Fix that urllib2 problem more thoroughly.
I was adding urllib2 to the list of deps, but not its deps.  The result was
that the initrd now doesn't contain the bisect module.  So now, take the same
mapping from before and add any found values in as something else to calculate
the deps of.  This looks like it's grabbing everything.  I hope.