Fix the SIGSEGV handler
Due to the way that python handles signals and threads, synchronous
signals--i.e., those that are delivered synchronously due to an attempt
to execute an instruction, such as SIGSEGV and SIGFPE--cannot be handled
in a multi-threaded program with signal handlers installed from python.
Move the signal handler deeper into a depths of isys.c and replace the
handleSegv handler with a method that installs the C signal handlers.