Always use iutil to start processes.
Replace direct uses of the subprocess module and os.system with our
iutil wrappers to ensure consistency in process starting. Use
execWithRedirect for things that just want to run a command to
completion (most things), execWithCapture for the realm command that
uses stdout, execConsole for a debug shell in the TUI, and startProgram
for anything that manages the Popen object itself (nm-connection-editor,
yelp) or does something complicated (vncconfig).