Fix an issue with bad NFS info specified in source spoke.
If an NFS repo was specified as 10.0.0.1/example/dir, it would be
interpreted as HTTP. Instead of returning if no ':' is detected in
the address, this just continues and lets the ValueError get raised
when the url can't be split and displays an error to the user that
there was a problem setting up the source repo. No more mix-up with
it being HTTP either.