Simplify iutil.execReadlines.
Remove the thread and queue between execReadlines and the process it
starts. Instead of reading child process output into an unbounded
buffer, have the child block until we are ready to read input, and kill
the child when we are done reading input. Return an iterator object
instead of a generator so that we can kill the child on __del__.

Add a test for whether the child is killed when the iterator is
unreferenced.