Adapt to corrected interpetation of logvol --percent.
logvol --percent and --size are now mutually exclusive. Percent does
the same thing as size, but in terms of a percentage of the vg's free
free space (after all requests with sizes have been allocated, and before
any of them have been grown). A logvol defined with percent can still be
growable. In fact, that's a good idea since the percentage allocation
does not make any effort to allocate leftover extents after a
conservative calculation of the percentage-based sizes.

Resolves: rhbz#1116435