Warn about uses of the string module.
The string module is about half-deprecated, so we need to be careful
about which parts of it we use. We don't use it too much, so warn about
any use and filter out the correct ones with pylint pragmas.

Remove the uses of string.split, since strings can do that themselves
nowadays and string.split is removed in python3, and string.letters,
since this is also removed in python3 and we actually meant
string.ascii_letters.