Allow cdrom-swapping when doing "inst.ks=cdrom[:...]"
Apparently this is a thing: some people do installs in secure
environments where they have no network and aren't allowed to use USB
devices; all they can use is optical media. So they have the install
disc, and another cdrom that contains their kickstart.

So: the user boots the install DVD with "inst.ks=cdrom[:/path/ks.cfg]",
and once udev notices the disc we run both fetch-kickstart-disk and
anaconda-diskroot¹.

Normally, anaconda-diskroot would mount the installer DVD here, but we
need to be able to eject it, so anaconda-diskroot instead ignores it and
quietly exits.

Meanwhile, fetch-kickstart-disk tries to grab the kickstart and fails
(because we're looking at the install DVD!), so it asks the user to
insert the CDROM containing the kickstart.

So: when the user ejects the DVD and inserts their kickstart CD,
fetch-kickstart-disk will fetch the kickstart and ask for the install
media back.

Finally, the user re-inserts the DVD, anaconda-diskroot mounts it, and
off we go!

¹I'm assuming the ordering is undefined. It might be predictable in
reality but it's probably better if we don't rely on that.

Resolves: rhbz#1122104