Avoid the possibility of size variables being unset
Check for missing sizes in the logvol and part commands. If unset in
logvol and not using --percent, raise an error. If unset in part, set
size to None and let blivet figure it out.

Remove the TypeError checks for the %d operation failing since
pykickstart already checks for integer validity.

Related: rhbz#1116435